Chore: rename fetch.ts -> Fetch.ts
